Harden review edges: 64-bit health fade, non-object health payloads, null charge_state
Addresses the Gemini review on PR #782:

- power_logic.h: compute the cycle-fade with a 64-bit intermediate so a
  corrupted NVS cycle count near UINT32_MAX clamps to the 60% floor
  instead of wrapping past the cap back to a "healthy" reading. Host
  test pins both overflow cases. WAP sketch copy re-synced.
- sensor.py: a health payload that is valid JSON but not an object
  (array/number) now routes to the existing except block instead of
  crashing the MQTT handler with an uncaught AttributeError.
- health_metrics.py: charge_state explicitly null no longer stringifies
  to "None" before the comparison (no behavior change -- "none" never
  matched charging/full -- just removes the gotcha).

+++ firmware/common/power/power_logic.h
@@ -129,8 +129,10 @@ inline uint8_t classify_hw_charge_state(uint16_t battery_mv,
 inline uint8_t health_pct_from_cycles(uint32_t charge_cycles,
                                       bool battery_present) {
   if (!battery_present) return 100;
-  uint32_t fade =
-      (charge_cycles * HEALTH_FADE_PCT_PER_RATED_LIFE) / HEALTH_RATED_CYCLES;
+  /* 64-bit intermediate: a corrupted NVS cycle count near UINT32_MAX
+   * must clamp to the floor, not wrap past the cap back to "healthy". */
+  uint64_t fade = ((uint64_t)charge_cycles * HEALTH_FADE_PCT_PER_RATED_LIFE) /
+                  HEALTH_RATED_CYCLES;
   if (fade > HEALTH_FADE_CAP_PCT) fade = HEALTH_FADE_CAP_PCT;
   return (uint8_t)(100 - fade);
 }
